// Backend/src/api/v1/admin/settings/enums.ts
export enum SettingGroup {
  GENERAL = "general",
  BRANDING = "branding",
  LOCALIZATION = "localization",
  REGIONAL = "regional",
  CUSTOMER_APP = "customer_app",
  CLIENT_APPLICATIONS = "client_applications",
}

export enum ApplicationType {
  ADMIN = "admin",
  CUSTOMER = "customer",
  DELIVERY = "delivery",
}

export enum AnnouncementStyle {
  INFO = "info",
  WARNING = "warning",
  EMERGENCY = "emergency",
}